[UPDATE] currentUser Util
